requirements.txt
beautifulsoup4
requests
pydantic
spotipy
python-dotenv
tqdm